added login functionality with firebase

+++ OneBook/app/src/main/java/com/example/onebook/Owner.java
@@ -2,14 +2,14 @@ package com.example.onebook;
 
 public class Owner extends User {
     
-    Owner(String username, String password, String email, long phone, int userID) {
+    Owner(String username, String password, String email, long phone, long userID) {
         super(username, password, email, phone, userID);
     }
     Owner(String username, String password, String email){
         super(username, password, email);
     }
 
-    public boolean lendBook(Book book, int userID) {
+    public boolean lendBook(Book book, long userID) {
         //code to be added
         return true;
     }
